server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
# Blockchain/Web3
web3==6.11.1
eth-account==0.9.0
python-dotenv==1.0.0
# AI/Numerical Verification (AMD ROCm compatible)
numpy==1.26.1
torch==2.1.0+cpu
# If AMD GPU acceleration is available:
# --extra-index-url https://download.pytorch.org/whl/rocm5.6
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
web3==6.11.1
eth-account==0.9.0
python-dotenv==1.0.0
# Android Logic
pure-python-adb==0.3.0.dev0
# AI/Numerical (AMD ROCm)
numpy==1.26.1
torch==2.1.0 --index-url https://download.pytorch.org/whl/cpu